+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.15 - 2024-12-01
+
+### Added
+- add a new OSC Control node `VrchDelayOscControlNode`
+
+### Updated
+
+- update `VrchAnyOSCControlNode` to support IS_CHANGED() based on OSC message value change
+- update `osc_control_nodes.md` documentation
+
 ## 1.0.14 - 2024-11-24
 
 ### Added